Refactor: Move message definitions to `dora-message` crate
First step towards versioning the message definitions indepedently.
