+++ apis/rust/node/Cargo.toml
@@ -12,6 +12,7 @@ tracing = ["dep:dora-tracing"]
 
 [dependencies]
 dora-core = { workspace = true }
+dora-message = { workspace = true }
 shared-memory-server = { workspace = true }
 eyre = "0.6.7"
 serde_yaml = "0.8.23"
