+++ binaries/cli/Cargo.toml
@@ -20,6 +20,7 @@ tracing = ["dep:dora-tracing"]
 clap = { version = "4.0.3", features = ["derive"] }
 eyre = "0.6.8"
 dora-core = { workspace = true }
+dora-message = { workspace = true }
 dora-node-api-c = { workspace = true }
 dora-operator-api-c = { workspace = true }
 serde = { version = "1.0.136", features = ["derive"] }
