+++ binaries/cli/src/logs.rs
@@ -1,5 +1,5 @@
 use communication_layer_request_reply::TcpRequestReplyConnection;
-use dora_core::topics::{ControlRequest, ControlRequestReply};
+use dora_message::{cli_to_coordinator::ControlRequest, coordinator_to_cli::ControlRequestReply};
 use eyre::{bail, Context, Result};
 use uuid::Uuid;
