+++ binaries/daemon/Cargo.toml
@@ -29,6 +29,7 @@ dora-download = { workspace = true }
 dora-tracing = { workspace = true, optional = true }
 dora-arrow-convert = { workspace = true }
 dora-node-api = { workspace = true }
+dora-message = { workspace = true }
 serde_yaml = "0.8.23"
 uuid = { version = "1.7", features = ["v7"] }
 futures = "0.3.25"
