+++ binaries/daemon/src/node_communication/tcp.rs
@@ -5,10 +5,9 @@ use crate::{
     socket_stream_utils::{socket_stream_receive, socket_stream_send},
     Event,
 };
-use dora_core::{
-    config::DataId,
-    daemon_messages::{DaemonReply, DaemonRequest, Timestamped},
-    message::uhlc::HLC,
+use dora_core::{config::DataId, uhlc::HLC};
+use dora_message::{
+    common::Timestamped, daemon_to_node::DaemonReply, node_to_daemon::DaemonRequest,
 };
 use eyre::Context;
 use tokio::{
