+++ binaries/daemon/src/node_communication/unix_domain.rs
@@ -1,9 +1,8 @@
 use std::{collections::BTreeMap, io::ErrorKind, sync::Arc};
 
-use dora_core::{
-    config::DataId,
-    daemon_messages::{DaemonReply, DaemonRequest, Timestamped},
-    message::uhlc::HLC,
+use dora_core::{config::DataId, uhlc::HLC};
+use dora_message::{
+    common::Timestamped, daemon_to_node::DaemonReply, node_to_daemon::DaemonRequest,
 };
 use eyre::Context;
 use tokio::{
