+++ libraries/core/Cargo.toml
@@ -15,11 +15,10 @@ serde_yaml = "0.9.11"
 once_cell = "1.13.0"
 which = "5.0.0"
 uuid = { version = "1.7", features = ["serde", "v7"] }
-dora-message = { workspace = true }
 tracing = "0.1"
 serde-with-expand-env = "1.1.0"
 tokio = { version = "1.24.1", features = ["fs", "process", "sync"] }
-aligned-vec = { version = "0.5.0", features = ["serde"] }
 schemars = "0.8.19"
 serde_json = "1.0.117"
 log = { version = "0.4.21", features = ["serde"] }
+uhlc = "0.5.1"
