+++ libraries/message/Cargo.toml
@@ -11,7 +11,11 @@ license.workspace = true
 
 [dependencies]
 arrow-data = { workspace = true }
-uhlc = "0.5.1"
 serde = { version = "1.0.136", features = ["derive"] }
 eyre = "0.6.8"
 arrow-schema = { workspace = true, features = ["serde"] }
+tokio = "1.39.2"
+dora-core = { workspace = true }
+uuid = { version = "1.7", features = ["serde", "v7"] }
+log = { version = "0.4.21", features = ["serde"] }
+aligned-vec = { version = "0.5.0", features = ["serde"] }
