+++ libraries/message/src/coordinator_to_cli.rs
@@ -1,0 +1,85 @@
+use std::collections::{BTreeMap, BTreeSet};
+
+use dora_core::config::NodeId;
+use dora_core::uhlc;
+use uuid::Uuid;
+
+pub use crate::common::LogMessage;
+pub use crate::common::{NodeError, NodeErrorCause, NodeExitStatus};
+
+#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
+pub enum ControlRequestReply {
+    Error(String),
+    CoordinatorStopped,
+    DataflowStarted { uuid: Uuid },
+    DataflowReloaded { uuid: Uuid },
+    DataflowStopped { uuid: Uuid, result: DataflowResult },
+    DataflowList(DataflowList),
+    DestroyOk,
+    DaemonConnected(bool),
+    ConnectedMachines(BTreeSet<String>),
+    Logs(Vec<u8>),
+}
+
+#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
+pub struct DataflowResult {
+    pub uuid: Uuid,
+    pub timestamp: uhlc::Timestamp,
+    pub node_results: BTreeMap<NodeId, Result<(), NodeError>>,
+}
+
+impl DataflowResult {
+    pub fn ok_empty(uuid: Uuid, timestamp: uhlc::Timestamp) -> Self {
+        Self {
+            uuid,
+            timestamp,
+            node_results: Default::default(),
+        }
+    }
+
+    pub fn is_ok(&self) -> bool {
+        self.node_results.values().all(|r| r.is_ok())
+    }
+}
+
+#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
+pub struct DataflowList(pub Vec<DataflowListEntry>);
+
+impl DataflowList {
+    pub fn get_active(&self) -> Vec<DataflowIdAndName> {
+        self.0
+            .iter()
+            .filter(|d| d.status == DataflowStatus::Running)
+            .map(|d| d.id.clone())
+            .collect()
+    }
+}
+
+#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
+pub struct DataflowListEntry {
+    pub id: DataflowIdAndName,
+    pub status: DataflowStatus,
+}
+
+#[derive(Debug, Clone, Copy, serde::Deserialize, serde::Serialize, PartialEq, Eq)]
+pub enum DataflowStatus {
+    Running,
+    Finished,
+    Failed,
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub struct DataflowIdAndName {
+    pub uuid: Uuid,
+    pub name: Option<String>,
+}
+
+impl std::fmt::Display for DataflowIdAndName {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        if let Some(name) = &self.name {
+            write!(f, "[{name}] {}", self.uuid)
+        } else {
+            write!(f, "[<unnamed>] {}", self.uuid)
+        }
+    }
+}
